refactor: clean up anthropic streaming helpers
