+++ Sources/AnthropicSession/AnthropicAdapter.swift
@@ -284,227 +284,4 @@ public actor AnthropicAdapter: Adapter {
 
     return toolCalls
   }
-
-  func executeToolCall(
-    _ toolCall: Transcript.ToolCall,
-    generatedTranscript: inout Transcript,
-    continuation: AsyncThrowingStream<AdapterUpdate, any Error>.Continuation,
-  ) async throws {
-    guard let tool = tools.first(where: { $0.name == toolCall.toolName }) else {
-      AgentLog.error(
-        GenerationError.unsupportedToolCalled(.init(toolName: toolCall.toolName)),
-        context: "tool_not_found",
-      )
-      throw GenerationError.unsupportedToolCalled(.init(toolName: toolCall.toolName))
-    }
-
-    do {
-      let output = try await callTool(tool, with: toolCall.arguments)
-
-      let toolOutputEntry = Transcript.ToolOutput(
-        id: toolCall.id,
-        callId: toolCall.callId,
-        toolName: toolCall.toolName,
-        segment: .structure(
-          Transcript.StructuredSegment(content: output),
-        ),
-        status: nil,
-      )
-
-      AgentLog.toolOutput(
-        name: tool.name,
-        callId: toolCall.callId,
-        outputJSONOrText: output.generatedContent.stableJsonString,
-      )
-
-      let transcriptEntry = Transcript.Entry.toolOutput(toolOutputEntry)
-      generatedTranscript.entries.append(transcriptEntry)
-      continuation.yield(.transcript(transcriptEntry))
-    } catch let toolRunRejection as ToolRunRejection {
-      let toolOutputEntry = Transcript.ToolOutput(
-        id: toolCall.id,
-        callId: toolCall.callId,
-        toolName: toolCall.toolName,
-        segment: .structure(
-          Transcript.StructuredSegment(content: toolRunRejection.generatedContent),
-        ),
-        status: nil,
-      )
-
-      AgentLog.toolOutput(
-        name: tool.name,
-        callId: toolCall.callId,
-        outputJSONOrText: toolRunRejection.generatedContent.stableJsonString,
-      )
-
-      let transcriptEntry = Transcript.Entry.toolOutput(toolOutputEntry)
-      generatedTranscript.entries.append(transcriptEntry)
-      continuation.yield(.transcript(transcriptEntry))
-    } catch {
-      AgentLog.error(error, context: "tool_call_failed_\(tool.name)")
-      throw GenerationError.toolExecutionFailed(toolName: tool.name, underlyingError: error)
-    }
-  }
-
-  func messageRequest(
-    including transcript: Transcript,
-    generating type: (some StructuredOutput).Type?,
-    using model: Model,
-    options: AnthropicGenerationOptions,
-    streamResponses: Bool,
-  ) throws -> MessageParameter {
-    let messages = try AnthropicMessageBuilder.messages(
-      from: transcript,
-      includeThinking: options.thinking != nil,
-    )
-
-    let systemPrompt: MessageParameter.System? = instructions.isEmpty
-      ? nil
-      : .text(instructions)
-
-    let toolChoice = try resolvedToolChoice(
-      for: type,
-      options: options,
-    )
-
-    let tools = try resolvedTools(
-      for: type,
-    )
-
-    return MessageParameter(
-      model: .other(model.rawValue),
-      messages: messages,
-      maxTokens: options.maxOutputTokens ?? 1024,
-      system: systemPrompt,
-      metadata: nil,
-      stopSequences: options.stopSequences,
-      stream: streamResponses,
-      temperature: options.temperature,
-      topK: options.topK,
-      topP: options.topP,
-      tools: tools,
-      toolChoice: toolChoice,
-      thinking: options.thinking,
-      container: nil,
-    )
-  }
-
-  private func resolvedToolChoice(
-    for type: (some StructuredOutput).Type?,
-    options: AnthropicGenerationOptions,
-  ) throws -> MessageParameter.ToolChoice? {
-    if let type {
-      guard let name = try structuredOutputToolName(for: type.name) else {
-        throw GenerationError.requestFailed(
-          reason: .invalidRequestConfiguration,
-          detail: "Structured output requires a tool name.",
-        )
-      }
-
-      return MessageParameter.ToolChoice(
-        type: .tool,
-        name: name,
-        disableParallelToolUse: true,
-      )
-    }
-
-    return options.toolChoice
-  }
-
-  private func resolvedTools(
-    for type: (some StructuredOutput).Type?,
-  ) throws -> [MessageParameter.Tool]? {
-    var resolved: [MessageParameter.Tool] = try tools.map { tool in
-      try .function(
-        name: tool.name,
-        description: tool.description,
-        inputSchema: tool.parameters.asAnthropicJSONSchema(),
-        cacheControl: nil,
-      )
-    }
-
-    if let type {
-      guard let name = try structuredOutputToolName(for: type.name) else {
-        throw GenerationError.requestFailed(
-          reason: .invalidRequestConfiguration,
-          detail: "Structured output requires a tool name.",
-        )
-      }
-
-      let schema = type.Schema.generationSchema
-      let jsonSchema = try schema.asAnthropicJSONSchema()
-
-      resolved.append(
-        .function(
-          name: name,
-          description: "SwiftAgent structured output: \(type.name)",
-          inputSchema: jsonSchema,
-          cacheControl: nil,
-        ),
-      )
-    }
-
-    return resolved.isEmpty ? nil : resolved
-  }
-
-  func structuredOutputToolName(
-    for typeName: String?,
-  ) throws -> String? {
-    guard let typeName else {
-      return nil
-    }
-
-    let baseName = "swiftagent_structured_output"
-
-    if tools.contains(where: { $0.name == baseName }) {
-      throw GenerationError.requestFailed(
-        reason: .invalidRequestConfiguration,
-        detail: "Tool name conflict with reserved tool \"\(baseName)\".",
-      )
-    }
-
-    _ = typeName
-    return baseName
-  }
-
-  func transcriptStatus(
-    from stopReason: String?,
-  ) -> Transcript.Status {
-    switch stopReason {
-    case "max_tokens":
-      .incomplete
-    default:
-      .completed
-    }
-  }
-
-  func tokenUsage(
-    from response: MessageResponse,
-  ) -> TokenUsage? {
-    tokenUsage(from: response.usage)
-  }
-
-  func tokenUsage(
-    from usage: MessageResponse.Usage,
-  ) -> TokenUsage? {
-    let inputTokens = usage.inputTokens
-    let outputTokens = usage.outputTokens
-    let totalTokens = inputTokens.map { $0 + outputTokens }
-
-    return TokenUsage(
-      inputTokens: inputTokens,
-      outputTokens: outputTokens,
-      totalTokens: totalTokens,
-      cachedTokens: usage.cacheReadInputTokens,
-      reasoningTokens: usage.thinkingTokens,
-    )
-  }
-
-  func callTool<T: FoundationModels.Tool>(
-    _ tool: T,
-    with generatedContent: GeneratedContent,
-  ) async throws -> T.Output where T.Output: ConvertibleToGeneratedContent {
-    let arguments = try T.Arguments(generatedContent)
-    return try await tool.call(arguments: arguments)
-  }
 }
